requirements.txt
Rpi.GPIO
ConfigParser
Flask==0.11
Flask-security